Everything: Change style to match Mac OS 9
This patch disables rounded corners, modifies the placement of
titlebar buttons and makes it all look a bit more like Mac OS 9

Beware that this is all pretty much hardcoded and messy like the
original source code is. My main plan is to improve this so I'm
not planning to use the System7 or Mac OS 8 themes at all.